backend/users/serializers.py
from rest_framework import serializers
from . import models

# For User List
class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.CustomUser
        fields = ('email', 'username', )

